Class refactoring
