+++ libcloudforensics/providers/gcp/internal/gke.py
@@ -14,57 +14,20 @@
 # limitations under the License.
 """Google Kubernetes Engine functionalities."""
 
-import abc
 from typing import TYPE_CHECKING, Any, Dict, List
 from kubernetes import client
 from kubernetes.config import kube_config
 from libcloudforensics.providers.gcp.internal import compute
 from libcloudforensics.providers.gcp.internal import common
+from libcloudforensics.providers.kubernetes import K8sResource, K8sSelector
 
 if TYPE_CHECKING:
   import googleapiclient
 
 
-class K8sSelector:
-  """Class to build K8s API selectors."""
-
-  class Component(abc.ABC):
-    """Component of the selector."""
-
-    @abc.abstractmethod
-    def ToString(self):
-      """Returns the component of the selector."""
-
-  class Node(Component):
-    """Component for running on a particular node."""
-
-    def __init__(self, node) -> None:
-      super().__init__()
-      self.node = node
-
-    def ToString(self):
-      return 'spec.nodeName={0:s}'.format(self.node)
-
-  class Running(Component):
-    """Component for a running pod."""
-
-    def ToString(self):
-      return 'status.phase!=Failed,status.phase!=Succeeded'
-
-
-  def __init__(self, *selectors):
-    self.selectors = selectors
-
-  def ToString(self):
-    """Builds the selector string to be passed to the K8s API."""
-    return ','.join(map(lambda s: s.ToString(), self.selectors))
-
-
 class GoogleKubernetesEngine:
-  """Class to call Google Kubernetes Engine (GKE) APIs.
+  """Base class for calling GKE APIs."""
 
-  https://cloud.google.com/kubernetes-engine/docs/reference/rest
-  """
   GKE_API_VERSION = 'v1'
 
   @staticmethod
@@ -80,8 +43,12 @@ class GoogleKubernetesEngine:
         'container', GoogleKubernetesEngine.GKE_API_VERSION)
 
 
-class GkeCluster(GoogleKubernetesEngine):
-  """Class facilitating GKE API and K8s API functions calls on a cluster."""
+class GkeResource(GoogleKubernetesEngine, K8sResource):
+  """Class to call GKE and Kubernetes APIs on a GKE resource.
+
+  https://cloud.google.com/kubernetes-engine/docs/reference/rest
+  https://kubernetes.io/docs/reference/
+  """
 
   def __init__(self, project_id, zone, cluster_id) -> None:
     super().__init__()
@@ -91,10 +58,10 @@ class GkeCluster(GoogleKubernetesEngine):
 
   @property
   def name(self):
-    """Property to retrieve the name of the cluster, for use in API calls.
+    """Name of the GKE cluster resource, for use in API calls.
 
     Returns:
-      str: Full name of the cluster.
+      str: Full name of the cluster resource.
     """
     return 'projects/{0:s}/locations/{1:s}/clusters/{2:s}'.format(
         self.project_id,
@@ -103,12 +70,6 @@ class GkeCluster(GoogleKubernetesEngine):
     )
 
   def _K8sApi(self) -> 'client.CoreV1Api':
-    """Creates an authenticated Kubernetes API client.
-
-    Returns:
-      kubernetes.client.CoreV1Api: An authenticated client to
-        the Kubernetes API server.
-    """
     # Retrieve cluster information via GKE API
     get = self.GetOperation()
     # Extract fields for kubeconfig
@@ -150,7 +111,7 @@ class GkeCluster(GoogleKubernetesEngine):
     return client.CoreV1Api(client.ApiClient(kubeconfig))
 
   def GetOperation(self) -> Dict[str, Any]:
-    """Get GKE API operation object for the GKE cluster.
+    """Get GKE API operation object for the GKE resource.
 
     Returns:
       Dict[str, Any]: GKE API response to 'get' operation for this
@@ -160,6 +121,10 @@ class GkeCluster(GoogleKubernetesEngine):
     request = clusters.get(name=self.name)
     response = request.execute()
     return response
+
+
+class GkeCluster(GkeResource):
+  """Class facilitating GKE API and K8s API functions calls on a cluster."""
 
   def GetInstances(self) -> List[compute.GoogleComputeInstance]:
     """Gets the GCE instances of the cluster".
@@ -175,14 +140,24 @@ class GkeCluster(GoogleKubernetesEngine):
       instances.append(instance)
     return instances
 
-  def GetRunningPods(self, node: str):
+
+class GkeNode(GkeResource):
+  """Class facilitating API functions calls on a GKE cluster's node."""
+
+  def __init__(
+      self, project_id: str, zone: str, cluster_id: str,
+      node_name: str) -> None:
+    super().__init__(project_id, zone, cluster_id)
+    self.node_name = node_name
+
+  def GetRunningPods(self):
     """Returns the pods running for a particular node"""
     selector = K8sSelector(
-      K8sSelector.Node(node),
-      K8sSelector.Running(),
+        K8sSelector.Node(self.node_name),
+        K8sSelector.Running(),
     )
     pods = self._K8sApi().list_pod_for_all_namespaces(
-      field_selector=selector.ToString())
+        field_selector=selector.ToString())
     for pod in pods.items:
       print(pod.metadata.name)
     return pods
